Use localmemcache for keeping the circuit state
[LocalMemCache](http://localmemcache.rubyforge.org/) uses memory mapped
files to share data across processes and keep it persistent on disk
while maintaining almost raw ruby Hash speed (losing about 10%) this
would be ideal for sharing the Circuitbreaker state per host.

Blockers: LocalMemCache does not support the increment API of moneta,
this is fixable by keeping a counter per process and summing the
counters on read, but we need to either do that in Circuitbox or patch
Moneta and/or localmemcache. Maybe we can even support atomic increment
in LocalMemCache?